server/requirements.txt
# [SECURITY] Patched Versions
fastapi>=0.109.1
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pdfplumber==0.10.3
# [SECURITY] Patched Buffer Overflow Vulnerability
Pillow>=10.3.0
pytesseract==0.3.10
# [SECURITY] Pinned version to prevent supply chain attacks
google-genai==1.0.0
pydantic==2.6.0
pydantic-settings==2.1.0
python-dotenv==1.0.1
firebase-admin==6.4.0
email-validator==2.1.0
aiosmtplib==3.0.1
jinja2==3.1.4
slowapi==0.1.9
structlog==24.1.0
sentry-sdk[fastapi]==1.40.0
